+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorViewModel.kt
@@ -52,6 +52,7 @@ internal class EpubNavigatorViewModel(
     val config: EpubNavigatorFragment.Configuration,
     initialPreferences: EpubPreferences,
     val layout: EpubLayout,
+    val listener: VisualNavigator.Listener?,
     private val defaults: EpubDefaults,
     baseUrl: String?,
     private val server: WebViewServer?,
@@ -76,7 +77,7 @@ internal class EpubNavigatorViewModel(
     }
 
     sealed class Event {
-        data class GoTo(val target: Link) : Event()
+        data class OpenInternalLink(val target: Link) : Event()
         data class OpenExternalLink(val url: Uri) : Event()
         /** Refreshes all the resources in the view pager. */
         object InvalidateViewPager : Event()
@@ -205,7 +206,9 @@ internal class EpubNavigatorViewModel(
         val href = url.toString()
         val link = internalLinkFromUrl(href)
         if (link != null) {
-            _events.send(Event.GoTo(link))
+            if (listener?.shouldJumpToLink(link) == true) {
+                _events.send(Event.OpenInternalLink(link))
+            }
         } else {
             _events.send(Event.OpenExternalLink(url))
         }
@@ -384,12 +387,13 @@ internal class EpubNavigatorViewModel(
             publication: Publication,
             baseUrl: String?,
             layout: EpubLayout,
+            listener: VisualNavigator.Listener?,
             defaults: EpubDefaults,
             config: EpubNavigatorFragment.Configuration,
             initialPreferences: EpubPreferences
         ) = createViewModelFactory {
             EpubNavigatorViewModel(
-                application, publication, config, initialPreferences, layout,
+                application, publication, config, initialPreferences, layout, listener,
                 defaults = defaults,
                 baseUrl = baseUrl,
                 server = if (baseUrl != null) null